chat-IA/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit-chat>=0.1.1
streamlit-option-menu>=0.3.6
plotly>=5.17.0
pandas>=2.1.0
